+++ urls4irl/templates/login.html
@@ -36,7 +36,7 @@
         </fieldset>
         <div class="form-group d-flex">
             {{ login_form.submit(class="btn btn-primary col-4 mr-4")}}
-            <a class="btn btn-primary col-7" href="{{url_for('register_user')}}">Register</a>
+            <a class="btn btn-primary col-7" href="{{url_for('users.register_user')}}">Register</a>
         </div>
     </form>
