+++ urls4irl/templates/register_user.html
@@ -81,7 +81,7 @@
 <div class="d-flex justify-content-center">
     <div class="border-top pt-3">
         <small class="text-muted">
-            <a href="{{url_for('login')}}">Already a user? Login</a>
+            <a href="{{url_for('users.login')}}">Already a user? Login</a>
         </small>
     </div>
 </div>
